fix: Automatic retry logic for unary requests
Minor refactoring to add a wrapper which can automatically retry when
unary requests fail with `UNAVAILABLE`. This also makes way for logic
like switching between different `etcd` server instances when one is
unreachable. There probably need to be a lot more changes before this is
good to go.

Also, there is a minor bit of `unsafe` code in the new
`ClientInner::wrap_unary_call` function. It `transmute`s the generic
`&mut GrpcClient` to a `&'a mut GrpcClient` so that it can be used in
the `call` function which requires a `&'a mut GrpcClient`. The issue is
that Tonic generated code uses `async fn foo(&mut self, ...)`, which is
bound to a specific lifetime, so that needs to be named, but it is not
clear how to give it a name and have it still be locally-bound. So here
it gets the name `'a` at the function level and then we pretend that the
`transmute` is safe. It probably is because Tonic generated code does
not keep a reference to `self` after the call completes.