feat: Adds `Record` and `AsKey` and `AsValue` converters
This replaces the `client::Entry` object with a new `record::Record`
structure. This is a monadic structure in three parts: key, value, and
metadata. All are templated, but they are generally expected to be two
`Vec<u8>`s and the `record::Metadata` type.

The `AsKey` and `AsValue` traits are meant for users to implement for
their custom key and value representations. There are implementors for
`[u8]`, `[u8; N]`, `Vec<u8>`, `str`, and `String`, as well as references
to these types, including `Box`es, `Rc`s, and `Arc`s. The return value
of the trait functions are `&[u8]`, which might be inefficient, since
the Tonic-generated bindings require a `Vec<u8>` for these (owned
structs like `String` and `Vec<u8>` can make this trivial). However, I
am hoping for a better story around in-place Protobuf encoding at some
point in the future.