feat: Create `AsRange` and use it in `list(...)`
The `AsRange` trait is like `AsKey`, except it covers Rust range
expressions. This changes the `Client::list` operation to be more
ergonomic: `client.list("a".."b")` instead of
`client.list().range("a".."b")`. Forcing the user to specify the range
at the `list` call also means whole-DB scans via `..` are explicit
instead of implicit by not calling the `range`/`prefix` function.

This also creates the `Prefix` function, which is useful for explicitly
specifying a prefix to scan for. Users probably do not have to use this,
as the `Client::list_prefix` is going to be more convenient.