feat: Adds leases API
This creates a basic API for creating and revoking leases. There is no
keep-alive API, as that would require doing streams. While I could just
use `tokio::sync` for this, that might be a big requirement to bring in.
That said, anyone using Tonic is using Tokio, so it probably does not
matter.